Form can now successfully submit user data which then can turn into a 'card' on the user UI (bookShelf).
The data from the form is passed to the Book consutrctor function using the 'new' keyword when the form is submitted. This creates a new book with that user form/input data and pushes this new Book object to the library array. The addNewBookToShelf() is called and takes the latest addition to the library array and adds it to a 'card' to be display to the user on the 'bookShelf'.