Form can now submit default values and requires user input for a title.
Logic added to the 'submit' event on the form, that will check to see if the input for pages and author is empty - if so, it will add 'Unknown' as the value to be submitted to for the new Book object. On the 'bookShelf' in that book's 'card' this will be seen as 'Unknown'.