Removed the addAllBooksToShelf() to clean up the code and reduce repetition. The addNewBookToShelf() can do this job with less lines.
The remove button will now confirm a user's decision before removing the card/book from the 'bookShelf'. Remove buttons now are associated with the Book instance that they were created with, by having the Book's index number stored in a property they have known as 'bookId'. This way we can reference this number and remove that item from the library array. 'bookCard's now have an id value associated with the index of the Book in the library too.