Fixed bug with the 'removeBookBtn' and 'bookCard' removal from calling the 'click' event on the 'removebBookBtn'.
The 'bookIndex' properties of the 'bookCard's and 'removeBookBtn's appropriately correlate to the new index of the library after removing the Book in question. By using the 'getElementsByClassName' selector instead of 'querySeletorAll' we now have a live node list that will actually update when we make changes to it. This now allows our indexing of the buttons and cards to correlate to the index of the list when we remove/delete these elements.